pyproject.toml
[tool.poetry]
name = "ftw-django-features"
version = "2026.2.0"
description = "A collection of features used in our Django-based web applications."
authors = ["4teamwork AG"]
readme = "README.md"
packages = [{include = "django_features"}]

[tool.poetry.dependencies]
python = ">=3.12, <4"
django-click = "^2"
django-configurations = "^2"
django-constance = "^4"
django-extensions = "^4"
django-filter = "^25"
django-linear-migrations = "^2"
django-modeltranslation = "^0.19"
djangorestframework = "^3"
django = ">=4.2, <6"
pluck = "^0.2"
python-dotenv = "^1"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
factory-boy = "^3.3.3"
flake8 = "7.3.0"
freezegun = "^1.5.5"
mypy = "^1.17.1"
psycopg2-binary = "^2.9.10"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
towncrier = "^25.8.0"
zest-releaser = "^9.6.2"
zestreleaser-towncrier = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
(
  /(
      \.git
    | \.venv
    | \.eggs
  )/
)
'''

[tool.isort]
force_alphabetical_sort_within_sections = true
force_single_line = true
from_first = false
line_length = 200
known_first_party = "app,django_features"
lines_after_imports = 2
skip_glob = ".venv,.eggs"

[tool.towncrier]
package = "django_features"
directory = "changes"
template = "towncrier.md"
filename = "CHANGELOG.md"
underlines = ["-", ""]

[[tool.towncrier.type]]
directory = "feature"
name = "New features:"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug fixes:"
showcontent = true

[[tool.towncrier.type]]
directory = "other"
name = "Other changes:"
showcontent = true

[tool.mypy]
follow_imports = "skip"
strict = true
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_subclassing_any = false
disallow_untyped_decorators = false
disallow_any_generics = false

[tool.pytest.ini_options]
addopts = "--nomigrations"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = "app"
DJANGO_SETTINGS_MODULE= "app.settings"
DJANGO_CONFIGURATION= "Development"
